First pass at locking
git-svn-id: svn+ssh://svn.bestpractical.com/svn/bps-public/rtir/branches/2.3-EXPERIMENTAL@7500 e417ac7c-1bcc-0310-8ffa-8f5827389a85

+++ lib/RT/IR/Ticket.pm
@@ -75,7 +75,18 @@ sub Lock {
 
 sub Unlock {
     my $ticket = shift;
+
+    my $lock = $ticket->RT::IR::Ticket::Locked();
+     return undef unless $lock;
+     return undef unless $lock->Content->{User} ==  $ticket->CurrentUser->id;
     $ticket->DeleteAttribute('RTIR_Lock');
 }
 
+
+sub BreakLock {
+    my $ticket = shift;
+    my $lock = $ticket->RT::IR::Ticket::Locked();
+     return undef unless $lock;
+    $ticket->DeleteAttribute('RTIR_Lock');
+}
 1;
